Refactor imports and restructure ventas module with models, schemas, routes, and service for improved organization
