+++ backend/app/db/create_db.py
@@ -1,5 +1,5 @@
-from app.db.database import Base, engine
-from app.models import venta  # noqa: F401
+from database import Base, engine
+from ventas.models import venta  # noqa: F401
 
 # importa tus modelos aquí
