+++ backend/app/db/database.py
@@ -1,4 +1,4 @@
-from app.core.config import settings
+from core.config import settings
 from sqlalchemy import create_engine
 from sqlalchemy.orm import declarative_base, sessionmaker
