+++ backend/app/main.py
@@ -1,7 +1,7 @@
-from app.api.routes import router
-from app.db.database import Base, engine
+from db.database import Base, engine
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
+from ventas.routes import router as ventas_router
 
 # Instancia principal de FastAPI
 app: FastAPI = FastAPI(title="Módulo de Registro de Ventas")
@@ -18,5 +18,5 @@ app.add_middleware(
 # Crear las tablas en la base de datos si no existen
 Base.metadata.create_all(bind=engine)
 
-# Incluir todas las rutas definidas en api/routes.py
-app.include_router(router)
+# Incluir todas las rutas de ventas
+app.include_router(ventas_router)
